fix tg->pg int->bigint, add dockerfile

+++ internal/ai/database.go
@@ -48,7 +48,7 @@ func (s *Service) CreateTables() error {
 	_, err = s.DBHandler.DB.Exec(`
 		CREATE TABLE IF NOT EXISTS auth (
 			id SERIAL PRIMARY KEY,
-			tg_user_id INT NOT NULL,
+			tg_user_id BIGINT NOT NULL,
 			auth_method INT REFERENCES auth_methods(id),
 			token TEXT NOT NULL
 		)`)
@@ -57,7 +57,7 @@ func (s *Service) CreateTables() error {
 	}
 	_, err = s.DBHandler.DB.Exec(`
 		CREATE TABLE IF NOT EXISTS ai_sessions (
-			tg_user_id INT PRIMARY KEY,
+			tg_user_id BIGINT PRIMARY KEY,
 			model TEXT,
 			endpoint INT REFERENCES endpoints(id)	
 		)`)
@@ -67,11 +67,11 @@ func (s *Service) CreateTables() error {
 	_, err = s.DBHandler.DB.Exec(`
 		CREATE TABLE IF NOT EXISTS chat_sessions (
 			id SERIAL PRIMARY KEY,
-			tg_user_id INT NOT NULL,
+			tg_user_id BIGINT NOT NULL,
 			model TEXT NOT NULL,
 			endpoint INT NOT NULL REFERENCES endpoints(id),
-			chat_id INT NOT NULL,
-			thread_id INT NOT NULL
+			chat_id BIGINT NOT NULL,
+			thread_id BIGINT NOT NULL
 		)`)
 	if err != nil {
 		return err
@@ -95,7 +95,7 @@ func (s *Service) CreateTables() error {
 		CREATE TABLE IF NOT EXISTS usage (
 			id BIGSERIAL PRIMARY KEY,
 			chat_session INT REFERENCES chat_sessions(id),
-			tg_user_id INT NOT NULL,
+			tg_user_id BIGINT NOT NULL,
 			completion_tokens INT,
 			prompt_tokens INT,
 			total_tokens INT,
@@ -127,8 +127,8 @@ func (s *Service) CreateTables() error {
 		CREATE OR REPLACE FUNCTION upsert_chat_session_and_usage(
 		    p_user_id BIGINT,
 		    p_endpoint_id INT,
-		    p_chat_id INT,
-		    p_thread_id INT,
+		    p_chat_id BIGINT,
+		    p_thread_id BIGINT,
 		    p_model TEXT,
 		    p_usage_completion_tokens INT,
 		    p_usage_prompt_tokens INT,
